update error message

+++ app/controllers/assign_shift_controller.rb
@@ -8,27 +8,43 @@ class AssignShiftController < ApplicationController
     @assignments = Assignment.where(course_id: params[:course_id]).where.not(teaching_assistant_id: TeachingAssistant.find_by(description: "dammy").id)
     @assigned_teaching_assistant = TeachingAssistant.where(id: @assignments.pluck(:teaching_assistant_id))
     @work_hour = WorkHour.where(assignment_id: Assignment.find_by(course_id: params[:course_id]))
-    @complete_assignment = WorkHour.joins(assignment: :teaching_assistant).where.not(assignments: { id: nil }).where.not(assignments: { teaching_assistant_id: TeachingAssistant.find_by(description: "dammy").id })
+    @complete_assignment = WorkHour.joins(assignment: :teaching_assistant).where.not(assignments: { id: nil }).where.not(assignments: { teaching_assistant_id: TeachingAssistant.find_by(description: "dammy").id }).where(assignments:{course_id:params[:course_id]})
   end
 
   def search
     number = params[:number]
     name = params[:name]
     @search_result = TeachingAssistant.where("name LIKE ? AND number LIKE ?", "%#{name}%","%#{number}%")
-    render json: @search_result
+
+    if @search_result.empty?
+      puts "検索結果がありません．"
+      render json: { error: "該当する学生が存在しません．" }, status: :unprocessable_entity
+    else
+      render json: @search_result
+    end
   end
+
 
 
   def add_TA
     selected_TA = params[:selected_options]
-    Assignment.add_ta(params[:course_id], selected_TA)
-    redirect_to request.referrer
+    if selected_TA.nil?
+      render json: { error: '対象者が選択されていません．' }, status: :bad_request
+    else
+      Assignment.add_ta(params[:course_id], selected_TA)
+      redirect_to request.referrer
+    end
   end
+
 
   def delete_TA
     selected_ta = params[:selected_items]
-    Assignment.delete_ta(selected_ta)
-    redirect_to request.referrer
+    if selected_ta.nil?
+      render json: { error: '削除対象者が選択されていません．' }, status: :unprocessable_entity
+    else
+      Assignment.delete_ta(selected_ta)
+      redirect_to request.referrer
+    end
   end
 
   def add_work_time
@@ -53,11 +69,12 @@ class AssignShiftController < ApplicationController
         work=WorkHour.find_by(id: work_time_id.to_i)
         work.destroy
       end
+      redirect_to request.referrer
     else
+      render json: { error: '削除する勤務時間が選択されていません．' }, status: :unprocessable_entity
       puts "勤務時間削除の対象が選択されていません．"
-
     end
-    redirect_to request.referrer
+
   end
 
   def add_assignment
@@ -71,11 +88,15 @@ class AssignShiftController < ApplicationController
 
   def delete_assgnment
     selected_assign = params[:selected_items]
-    selected_assign.each do |work_time_id|
-      work=WorkHour.find_by(id: work_time_id.to_i)
-      work.destroy
+    if selected_assign.nil?
+      render json: { error: '削除するシフトが選択されていません．' }, status: :unprocessable_entity
+    else
+      selected_assign.each do |work_time_id|
+        work=WorkHour.find_by(id: work_time_id.to_i)
+        work.destroy
+      end
+      redirect_to request.referrer
     end
-    redirect_to request.referrer
   end
